requirements.txt
# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# Cryptography
cryptography>=38.0.0

# GUI
PyQt5>=5.15.0

# Scientific computing
scipy>=1.7.0

# Optional: Machine Learning
# scikit-learn>=1.0.0
# tensorflow>=2.8.0

# Optional: Audio/Video processing
# pydub>=0.25.1
# moviepy>=1.0.3
argon2-cffi>=21.3.0